Removed cmp confirm select
Removed nvim-cmp completion tools's setting for automatically
selection the first item in the list if none are selected.

I prefer my Enter key to behave the same regardless of a menu being
up or not. Now I have to explicitly press <Tab> or <C-n> a few times
to select an item and complete it as such. That's the way I've been
doing it subconciously anyways.